Change auth, each user has a collection

+++ api/controllers/food-product.js
@@ -1,9 +1,10 @@
 const Product = require('../models/food-product')
 const mongoose = require('mongoose')
+const User = require('../models/user')
 
 exports.products_get_all = (req, res, next) => {
-	Product.find()
-		.select('_id name brand category img qty unit productQty expiry tags')
+	Product.find({ userId: req.userData.userId })
+		.select('_id userId name brand category img qty unit productQty expiry tags')
 		.exec()
 		.then(prods => {
 			const response = {
@@ -22,16 +23,15 @@ exports.products_get_all = (req, res, next) => {
 }
 
 exports.product_create = (req, res, next) => {
-	console.log(req.file)
 	const product = new Product({
 		_id: new mongoose.Types.ObjectId(),
+		userId: req.userData.userId,
 		name: req.body.name,
 		brand: req.body.brand,
 		category: req.body.category,
 		img: req.file.path,
 		qty: req.body.qty,
 		unit: req.body.unit,
-		productQty: req.body.productQty,
 		productQty: req.body.productQty,
 		expiry: req.body.expiry,
 		tags: req.body.tags
@@ -44,6 +44,7 @@ exports.product_create = (req, res, next) => {
 				message: 'Food product succesfully saved',
 				createdProduct: {
 					_id: result._id,
+					userId: req.userData.userId,
 					name: result.name,
 					brand: result.brand,
 					category: result.category,
@@ -66,8 +67,8 @@ exports.product_create = (req, res, next) => {
 
 exports.product_get = (req, res, next) => {
 	const id = req.params.productId
-	Product.findById(id)
-		.select('_id name brand category img qty unit productQty expiry tags')
+	Product.findOne({ _id: id, userId: req.userData.userId })
+		.select('_id userId name brand category img qty unit productQty expiry tags')
 		.exec()
 		.then(prod => {
 			console.log(prod)
@@ -87,13 +88,15 @@ exports.product_get = (req, res, next) => {
 
 exports.product_update = (req, res, next) => {
 	const id = req.params.productId
+	console.log(req.body)
 	// Get properties set in the body
 	const updateOps = {}
+
 	for (const ops of req.body) {
 		updateOps[ops.propName] = ops.value
 	}
 	// Update only the properties defined
-	Product.updateOne({ _id: id }, { $set: updateOps })
+	Product.updateOne({ _id: id, userId: req.userData.userId }, { $set: updateOps })
 		.exec()
 		.then(result => {
 			console.log(result)
@@ -111,17 +114,27 @@ exports.product_update = (req, res, next) => {
 
 exports.product_delete = (req, res, next) => {
 	const id = req.params.productId
-	Product.deleteOne({ _id: id })
+	Product.findOne({ _id: id, userId: req.userData.userId })
 		.exec()
 		.then(result => {
-			res.status(200).json({
-				message: 'Product ' + id + ' succesfully deleted'
-			})
+			if (!result) {
+				return res.status(500).json({
+					error: error
+				})
+			}
+			Product.deleteOne({ _id: id, userId: req.userData.userId })
+				.exec()
+				.then(result => {
+					res.status(200).json({
+						message: 'Product ' + id + ' succesfully deleted'
+					})
+				})
+				.catch(err => {
+					console.log(err)
+					res.status(500).json({
+						error: err
+					})
+				})
 		})
-		.catch(err => {
-			console.log(err)
-			res.status(500).json({
-				error: err
-			})
-		})
+		.catch()
 }
